requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
pydantic==2.6.3
httpx==0.27.0
python-dotenv==1.0.1
google-genai==0.3.0
twilio==9.0.2
slowapi==0.1.9
python-jose[cryptography]==3.3.0
tenacity==8.2.3
structlog==24.1.0
